requirements.txt
aiodns==3.2.0
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.2
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
annotated-types==0.7.0
anthropic==0.30.0
anyio==4.4.0
archspec @ file:///croot/archspec_1709217642129/work
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
boto3==1.34.136
botocore==1.34.136
Brotli @ file:///work/perseverance-python-buildout/croot/brotli-split_1698805593785/work
cachetools==5.4.0
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama==0.4.6
conda @ file:///croot/conda_1710772050586/work
conda-content-trust @ file:///work/perseverance-python-buildout/croot/conda-content-trust_1698882886606/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///work/perseverance-python-buildout/croot/conda-package-handling_1698851267218/work
conda_package_streaming @ file:///work/perseverance-python-buildout/croot/conda-package-streaming_1698847176583/work
cryptography @ file:///croot/cryptography_1710350347627/work
distro @ file:///work/perseverance-python-buildout/croot/distro_1701732366176/work
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-DB==0.4.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.6.1
google-auth==2.32.0
google-auth-oauthlib==1.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna @ file:///work/perseverance-python-buildout/croot/idna_1698845632828/work
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
jsonpatch @ file:///croot/jsonpatch_1710807507480/work
jsonpointer==2.1
libmambapy @ file:///croot/mamba-split_1712091911343/work/libmambapy
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1706732933928/work
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
openai==1.32.0
orjson==3.10.5
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.2
pandass==1.11.4
paramiko==3.4.0
pfzy==0.3.4
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
prettytable==3.10.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycares==4.4.0
pycosat @ file:///work/perseverance-python-buildout/croot/pycosat_1698863456259/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.7.3
pydantic-extra-types==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests @ file:///croot/requests_1707355572290/work
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
ruamel.yaml @ file:///work/perseverance-python-buildout/croot/ruamel.yaml_1698863605521/work
runpod==1.6.2
s3transfer==0.10.2
setuptools==68.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
starlette==0.37.2
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.5
tqdm @ file:///work/perseverance-python-buildout/croot/tqdm_1701735729845/work
tqdm-loggable==0.2
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typer==0.12.3
typing_extensions==4.12.1
tzdata==2024.1
ujson==5.10.0
urllib3 @ file:///croot/urllib3_1707770551213/work
uvicorn==0.30.1
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wheel==0.41.2
yarl==1.9.4
zstandard @ file:///work/perseverance-python-buildout/croot/zstandard_1698847073368/work
